Add function initVertexData to a Quad
Having this will open the opportunity to change the way to initialize
vertex data by overriding it. It is useful when creating a tile like or
particle like quad. For an example, we could use a global vertex data
instead of creating a new one for every quad that make it faster and
lighter.