Replacing multiple eventListeners with one
Hi!
I am proposing a small optimization change to the demo source where instead of adding an eventListener for each button of the main menu, only one eventListener is added to the 'Game.as' constructor. Also I am adding a condition to the onButtonTriggered method to deal with events generated by the back button in the 'Scene.as' class. This will also allow to simplify the 'Scene.as' class, by removing its own onBackButtonTriggered method. (I am going to edit Scene.as too, to show how).