refactoring once more: re-added contextData, but now it's bound to the stage3D instance
That way, the context data survives a context loss, and we've got the best of both worlds.
Actually, I had prefered 'Starling.instanceData', but to make that work when two Starling instances share the same context AND need to handle a context loss -- would have required Starling to be able to dispatch weak events (which is not possible ATM). Otherwise, memory issues would pop up if poeple forget to call 'dispose' on their display objects. Long story ...