more TouchProcessor refactoring
I made many of the properties of the "Touch" object writeable; this is required for TouchProcessor subclasses to create / modify their own touch objects. Furthermore, I removed the "updated" property of Touches and instead pass only the updated touches to the "processTouches" method.

The reasons for these changes can be found here:
https://github.com/PrimaryFeather/Starling-Framework/commit/a83d3537eff90b13bd9936bb4ab48d8b61cb8779