reverted VertexData from ByteArray to Vector - temporarily
The problem with ByteArrays is that each instance takes up 4k of memory. Until this is either changed in AIR or I have created a workaround, I have to switch back to Vector. I don't want to delay Starling 1.4 any longer, so I will publish it like that and tackle ByteArrays in 1.5.