Change to use a single QuadBatch
During development I found memory issues in this class due to the amount of QuadBatches that were being created with the associated Index and Vertex buffers. I have changed this to only use a single QuadBatch as there is no need to have multiple QuadBatches as the QuadBatch itself will grow its own index and vertex buffers as needed and keep them to the largest used size. So it is wasteful to have extra QuadBatches for this purpose. I have not noticed any performance decrease with this method, in fact there might even be a very small increase in performance.

I also added a function to dispose of the QuadBatch (disposeQuadBatch()) that is determined by the user to free the large index and vertex buffers that can accumulate in the QuadBatch.