fixed Image.getBounds and Quad.getBounds for non-zero upper left vertex case.
Texture's frame could be used to align images that is a rasterized
version of a flash symbol that have a custom pivot (in the center of an
image for example). In order to get correct bounds (and therefore
hitTest) one has to consider upper left vertex of an Image (Quad).