now prefering RectangleTextures even for POT textures (closes #409)
Because of Stage3D-internals, POT-Textures use up more memory as necessary (they always allocate mipmaps, whether you need or not). RectangleTextures do not suffer from this problem, so it makes sense to default to RectangleTextures all the time.

The downside: after it has been created, you cannot change the "repeat" property of a texture any more. The repeat setting must now be chosen when you create the texture. For this reason, all "Texture.from..." methods now contain a new "repeat" property.

Thus, this change breaks the API at this point (although I don't expect the repeat property to be in use very often).