replaced 'RenderUtil' with 'Painter' and 'RenderState'; added 'Pool' for basic object pooling
Over time, the old "RenderUtil" class had grown in size and had become rather confusing to use. The new setup is much cleaner: it splits up the code in two parts:

* a "RenderState" containing properties that are frequently changed, most importantly the render matrices and context settings.
* a "Painter" for general drawing operations, and for managing the state (using a stack). The stack makes it possible to write render code without the fear of side effects.

Furthermore, the Painter is publicly accessibly via a property on the Starling class — acting as a logical partner of the "Juggler" (the juggler animates, the painter draws). That way, you can access (and change!) the current render state even outside the render loop, e.g. when drawing to textures.