removed 'repeat' properties and method arguments from Texture classes
Most of the textures you use in Starling are being loaded from an atlas, so they do not support "repeat" anyway. This always made this property rather useless, so I decided to remove it altogether. Instead, there will be an equivalent of Feathers' TiledImage class, which works with all kinds of textures.

Of course, it's still possible to customize wrapping in custom display objects. (I might add a 'repeat' property to the MeshEffect class in the future.)