added 'BatchProcessor' and using it for all rendering
This class can work with different types of "IMeshBatch" implementations, making it possible to use different rendering code, depending on the object. However, it turned out that the use of interfaces is very slow on iOS (with the AOT compiler), so it's unsure if the implementation can stay like this.