added automatic render cache, providing much more efficient rendering of static parts of the display list
When an object does not change its position (alpha, ...) on the stage, its VertexData can now be copied directly from the previous frame. That works not only for single objects, but for containers with an arbitrary size. The only requirement is that self-rendering classes now call "setRequiresRedraw" to indicate that they have changed. Classes that do not support the cache (like Sprite3D) have to make this call every frame.

This commit also removes the "clipRect" property. The plan is instead to have "mask" automatically recognize rectangular clipping and use the scissor rectangle when possible. That will be added in a future commit.

Furthermore, this commit renders filters unusable. The complete filter API will be overhauled for Starling 2.0, anyway, and this problem will be fixed in this process.