converted all member variables using the m-prefix to underscores
The underscore prefix is more widely used among AS3 programmers, so I switched to that convention in Starling 2. All the new source files created so far were created with that convention already, but I wanted to delay the change for the existing classes, to remain being able to cherry-pick from the 1.x branch.

If anyone wants to make the same change to his/her projects, here is the script that made all those changes:
https://gist.github.com/PrimaryFeather/ba8b2eca965ac6f8baad