added new 'TextFormat' class and using it in TextField and BitmapFont; replaced 'HAlign' and 'VAlign' with one single 'Align' class
While it had advantages to be able to set the font properties directly on the TextField, it also had some ugly side-effects, e.g. the long list of arguments in some BitmapFont methods, and that it was difficult to predefine some formats and use them throughout a project.

Different to the implementation in classic Flash, though, you can change the format of a TextField on the fly, without having to constantly reassign the TextFormat for the changes to show up. This always bugged me in the past ...