fixed that render cache showed old content in certain situations
There was a problem with the render cache that didn't pop up before. The "BatchToken" instances in each display object store the range of vertex- and index-data in the previous BatchProcessor, so that they can be re-used if the object has not changed. However, even though this information is only valid for one frame (since the batch processors change constantly), I didn't store the frame ID in which this data was collected. This could lead to rendering of old content.

I noticed this in the benchmark scene after changing the buttons to "Scale9"-style. Hovering over the "back"-button would lead the button's background to change; this would then cause the button's text to render from the cache, but its token had been created before any benchmark objects. Thus, hovering over the button suddenly made new, static benchmark objects appear.