removed minimal non-zero alpha value from VertexData colors (closes #793)
The idea behind that was that all color data is lost once the alpha value reaches zero. However, with the current 8 bit resolution, this does not make much sense: the minimum alpha value would still mean that there are only 5 possible values per other channel, which is too little to be useful; and even with this value, an otherwise invisible object becomes slightly visible.

Thus, it's better to just be honest about it and let developers know that PMA has that effect on color accuracy.