refactoring of some RenderState and Painter procedures
Now, the RenderState reports to the Painter when the state changes in a way that requires a new draw call. E.g. a change of the render target will trigger a draw call right away; previously, this had to be done manually.

In the course of this change, I noticed that the BlendMode should be handled in the same way. Thus, MeshBatches change / check the blend mode no longer, but rely on the Painter's state logic instead. This seems to be much cleaner.