added the completely rewritten Fragment Filter API
The new filter API makes heavy use of the new rendering infrastructure of Starling 2, moving much of the heavy lifting to "Effect" subclasses. Filters can now use other filters for render passes; e.g. the DropShadowFilter is actually nothing more than a combination of the BlurFilter and the new CompositeFilter. The latter may combine up to four input textures into a single one, and is really only meant to be used by other filters.

The classes are not yet complete, e.g. the TexturePool (responsible for managing the textures) is just a stub right now, but all filters are working and are already automatically cached with Starling's standard render cache. What's also still missing is the new "FilterChain" which will allow a simple way to add multiple filters to one display object.