now deactivating render cache while drawing the filter's target object
The problem: all of the drawn objects updated their tokens even though the render cache is rewound later. So it could happen that it's attempted to redraw them from the cache, which would not contain the referenced batches any longer.

This happens e.g. when applying a filter over the root object of the Demo project, when any non-batchable text fields are visible.