changed fragment filters to be drawn in stage coordinates again
It turned out that it's impossible to render some nested Sprite3D content when using the local coordinate system; thus, I had to return to the Starling 1.x way of rendering filters in stage coordinates. The advantage is that I can now clip all filter content to the stage bounds, avoiding any unnecessary rendering. The disadvantage is that some filters that need to work in the local coordinate system of an object are more difficult to implement.