replaced MeshStyle's 'copyVertexData' and 'copyIndexData' with 'batchVertexData' and 'batchIndexData'
Note that it's not only the name that has changed, but also the type of the first parameter. We're no longer working directly with the vertex data, but with the target style instead. The reason for this is that you sometimes need access to the other style's data in the batching process; the vertex data alone is not always sufficient.

I realized that when making an experimental "MultiTextureStyle"; a style that allows to batch meshes with different textures together. This API change made that possible.