fixed alignment problems with HTML text
Starling 2 tries to use the minimal possible bitmap size when rendering TrueType text into a texture. However, with HTML text, that does not work, since it supports overriding alignments. The Flash TextField only informs me about the width and height of the text, but not to its position, so I can't know the actual text bounds when using HTML text.

Thus, I removed that optimization for HTML text.