making sure that ByteArray inside VertexData instance is never made smaller until "trim" is called
Otherwise, the render cache's memory is re-allocated every frame, which is bad for performance.

Note: This is the last commit that uses ByteArrays; the bad performance of ByteArray read/write operations forces me to switch back to 'Vector.<Number>'. This commit was just made so that I don't forget to make this change if I ever move back to ByteArrays.