rewrote 'scale9Grid' setup, fixing a bug and making sure that basic quad layout stays intact
Previously, I used the minimum number of vertices; however, this disrupts the basic quad layout of the IndexData on batching, which means that all subsquent quads in the same batch will be rendered much slower. The new version uploads more vertices, but keeps quad layout intact, which yields a better performance.

Furthermore, I realized that a certain scenario of Scale9Grids did not work: when using a texture with a frame and then using the complete width or height of the texture frame for the scale9Grid, the output was wrong.

I hope I never have to touch this method again! The combination of Scale9Grid and frame rectangle makes it really *extremely* painful to get all possible combinations right ...