added 'Starling.skipUnchangedFrames' property (closes #807)
This prevents any rendering if the stage (or any of its children) hasn't changed since the last frame, which is great for power consumption. However, this will also prevent VideoTextures from animating; I'd need to find out if they are present in the display list for this to work. The same is true for RenderTextures, but here it would be easy to solve (since I could simply set the stage 'dirty' in their 'draw' method).

In any case, until this is solved, I decided to disable this new feature by default. That way, it's more likely that people read the documentation about this property when they manually enable it (and don't run into any problems if they don't).

It's also interesting to note that on mobile, and on some Desktop environments in "baselineConstrained" profile, the classic display list is only updated when "context.present()" is called. This had to be put into consideration, as well.