added filter caching again
Just like in Starling 1.x, you can now call "cache()" to store the filter output and reuse it in the following frames. That's only possible for 2D objects, though: when trying to cache a filter on a Sprite3D (or a child of a Sprite3D), the operation will silently be ignored. (For this to work, we'd need a "Mesh3D", i.e. a mesh with vertices in a 3D space.)