added 'resolution' property to filter again
Just like in Starling 1.x, you can now set the resolution of a filter. That's especially important for the blur filter, since it generally works well with a lower resolution, providing a higher blur at lower processing time.

What was a little tricky here is that, different to Starling 1.x, DropShadow- and GlowFilter are not drawn behind the filtered object, but are composited into one texture. Thus, I first had to teach the CompositeFilter how to work with different texture scales.