changed how to handle objects that do not support the render cache
Now, those objects are supposed to simply call "painter.excludeFromCache()" in their render method. This will exclude them from being rendered from cache in the next frame. This is not only simpler to use, but is has one important advantage: the object is not marked as "requiresRedraw", and so Starling's "skipUnchangedFrames" policy is not disrupted.

Previously, when you had e.g. a masked object on the screen (even when stationary), the frame could not be skipped. Now, that's finally possible, which is great for Feathers-based apps (which use masks e.g. for lists).