now drawing the last fragment filter pass directly to the back buffer, if that makes sense
As long as the object is non-static (its "requiresRedraw" property returns true), the last filter pass is now rendered directly to the back buffer. That circumvents the render cache, but when the object is moving, it wouldn't be any use, anyway. Thus, it's better to spare that draw call. As soon as the object becomes static, the old behavior is restored.

Unfortunately, this logic fails when the target object contains a child that does not support the render cache. In those cases, the behavior can be manually tweeked via "filter.alwaysDrawToBackBuffer".

Beware: in this commit, I had to rename "ITexturePool" to "IFilterHelper", because the helper object now has more duties than just acquiring and releasing textures.