reverted default of 'MeshBatch.batchable' to 'false'
I activated this while being overjoyed about the introduction of the render cache. However, when people use the MeshBatch class, they normally use it to render a huge number of quads! And for that case, it's just a bad idea to do any batching. So I reverted this to the behavior of Starling 1.