fixed error that occurred when drawing objects with a filter to a RenderTexture (closes #810)
In principal, this now works: you can draw an object that has a filter attached to it to a RenderTexture. Unfortunately, though, you might get a wrong result when passing a custom matrix to the "draw" parameter:

    renderTexture.draw(object, matrix);

That's because the filter will always use the orientation of the object (its transformationMatrix) for its calculations; ideally, it should use the modelviewMatrix instead. However, the current API does not support this; I'd need a "getBounds" method that takes a matrix instead of an object as parameter. This would require a major API change, though, and it's too late for that in Starling 2. It's not critical, so this can definitely wait; but it needs to be considered for future updates.