made 'DisplayObject.setParent' starling_internal (instead of just 'internal')
The reason for this change: when developers want to make an object part of a container, but not let it show up via "getChild", it's necessary to still let the child know about its (somewhat hidden) parent-relationship. That way, things like "setRequiresRedraw" and "getTransformationMatrix" will work as designed. It's definitely something to be avoided by most developers, but in some situations it does make sense; thus the accessibility change.

Here's the discussion that lead to this change:
http://forum.starling-framework.org/topic/starling-20-skipunchangedframestrue-causes-alpha-tween-on-image-to-be-ignored