now dispatching initial CONTEXT3D_CREATE even when context is shared
Yes, it's not entirely accurate: with a shared context, the context is not created at that time by Starling, but somewhere else. However, anyone who's sharing the context should be aware of that anyway, and it's still useful to have an event that happens right before the root object is initialized.

Besides, it worked the same in Starling 1.x and it's not necessary to break anything here.