enhanced render cache logic: cleaner code, better reliability
Since the introduction of the render cache, I often had to add fixes for small problems. Some of the affected areas:

* Drawing into a RenderTexture: object was taken from the cache, which is not always desired in this case (#881)
* When drawing filters, only the last step needs to be cached. The process was a hack before, requiring to reverse the position inside the render cache and temporarily setting the frameID to zero.
* Masks caused several issues (e.g. #870), the fixes of which made the code of `DisplayObject.setRequiresRedraw` and `DisplayObjectContainer.render` more complicated than what felt necessary.

For this reason, the cache can now be temporarily disabled via `Painter.cacheEnabled`. When it's disabled, nothing will be drawn from cache, and nothing will be written to the cache. That's now the default for masks, `RenderTexture.draw`, and generally outside Starling's render code.

This makes the code much cleaner, and will hopefully avoid some problems in the future.