More data consistency
The "texture" is a through-constructor initialized immutable field. The case of which the two arguments are <code>null</code> is the "default" approach to utilize the "minimal" embedded font, and the "typical" case is when the two arguments are provided correctly. The <code>else</code> statement is to assert for data consistency and considering the current implementation, there is no chance to "set" the "texture" nor the "XML" later, so both of them need to be provided "not <code>null</code>" initially. However, only the specific case of the <code>texture != null && fontXml == null</code> is evaluated leaving the case of <code>texture == null && fontXml != null</code> unregarded! To compensate the issue, change the criteria to <code>texture == null || fontXml == null</code> to regard the both of the "wrong" cases :)