now updating cached texture base when MeshStyle is copied
The cached base reference is only used to find out if a mesh can be batched with another one.
When the texture is restored after a context loss, the base changes, but the reference is not updated.
This breaks batching, but doesn't cause any other rendering issues.

Ideally, the reference should be updated when the base texture changes, but there is no such mechanism yet.
Instead, I'm now simply updating it in the `copyFrom` method, since that is called regularly (without being overly expensive).
For the time being, this should do the trick.

This came up in this forum thread:
http://forum.starling-framework.org/topic/bug-with-context-restore