Employing the `numSides` parameter for elliptical gen methods
The useful parameter of `numSides` in the constructor of the internal class of `Ellipse` is implemented and available, why not allow the users to benefit the advantages of the parameter?! As the class provides a resolution-independent concept, one may need to have more control on the sides' count, e.g. where the shape may be scaled in a way the imperfection of the to-be elliptical curve is undesirable, and the issue could've been solved, if there were more sides on the shape!