now allowing to specify the alpha value with which the context is cleared
This feature is a little hidden: 'stage.color' is now interpreted as ARGB.
The issue came up because the particle system extension showed strange results when the particles are placed directly on top of the stage.
This can now be solved by changing the alpha value of the stage to '1.0'.