added workaround for problems when using masks and render textures simultaneously (closes #992)
For some reason, Stage3D doesn't use the stencil reference value correctly on render textures while a mask is used elsewhere.
This problem only popped up after the default reference value was changed from '0' to '127' (for inverted mask support).

It seems to me that this is actually a bug in Stage3D, but I have yet to pinpoint the exact problem.
Until then, the workaround is to ignore all stencil testing at the beginning of mask drawing.
(If there is actually a masked object being drawn, stencil actions are going to be reset, anyway.)