added 'quality' property to Blur-, DropShadow-, and GlowFilter (refs #1007)
On the DropShadow- and GlowFilters, the quality now replaces the `resolution` argument.
The result is similarly blurred, but fixes some problems that showed up in lower resolutions.

The blur effect always needs to take the current pixel size into account, and if it is used as part of a composite filter, different passes might use different pixels sizes - leading to artifacts or the blur being not uniform (stronger horizontal than vertical).