now fixed asset manager's memory issues at their roots
The reason that some memory was left in memory: apparently, closures don't keep references to just the objects they _need_, but to all that are available (i.e. the parent function's stack).
Extensive tests with Adobe Scout strongly indicate that.
So I'm now manually setting big references (ByteArray, BitmapData) to null in the parent method; that fixes the issue.