now using two separate sets of static (helper) matrix and point objects in DisplayObjectContainer
See this forum report: https://forum.starling-framework.org/d/22337-issue-with-button-in-starling-26

With the new "ButtonBehavior" being used by the Button class, hit tests on the button suddenly didn't work correctly any longer if there was an empty container or canvas added to its overlay.

Tracking it down, I noticed that in "DisplayObjectContainer.hitTest" the static helper matrix did not contain the correct matrix immediately after copying from "child.transformationMatrix". That's extremely strange, because how could the object change between those two lines?

However, it must be a side effect of the matrix (and point) objects being static helper objects that are shared with (1) recursive calls to the same method and (2) with the "hitTest" method. So I ruled out (2) by creating two separate helper objects for the latter. This fixed the issue. (But I still can't put my finger exactly on where the side effect occurred.)