fixed that context settings were not correctly restored after a context loss
The painter always stores the context's current values for blend mode, culling, depth mask, and depth test.
That way, when e.g. the blend mode is set to the value it's currently set to, anyway, this can be skipped.

However, those "shadow"-values were not reset after a context loss, meaning that the current context configuration was actually undefined.
With this change, the values are forcibly synced with the context in its restoration process.

Here's the report that lead to this finding:
https://forum.starling-framework.org/d/22765-rendertexture-onrestore-reversed