Update ConcretePotTexture.as
When uploading a 2048x2048 texture, all smaller mipmap levels were then uploading as 1024x1024, because that was the only BitmapData size that ever got created in the whole mipmap upload process. I changed it so now the deeper levels upload as BitmapData objects that are sized 512x512, 256x256, etc. Was it correct for me to change this, or am I making a mistake?

Also I am curious, why doesn't Starling support async uploads for mipmaps, is it due to hardware limitations or something else?