Update ConcretePotTexture.as
When uploading a 2048x2048 texture, all smaller mipmap levels were then uploading as 1024x1024, because that was the only BitmapData size that ever got created in the whole mipmap upload process. I changed it so now the deeper levels upload as BitmapData objects that are sized 512x512, 256x256, etc. Was it correct for me to do this?

More importantly, on older hardware like an HTC One, it seems like it can take around 230 milliseconds just for the CPU to scale a 2048x2048 BitmapData to any other scale with smoothing enabled. That means when scaling it to all 12 mipmap levels, it was taking 2760 milliseconds, which is almost 3 full seconds. So if you have 6 textures that are 2048x2048 which must load at launch, it can take 18 seconds just for the app to finish opening. I have changed the mipmap drawing process so that instead of scaling the full 2048x2048 texture down to each mipmap level each time, instead it scales by 50% at each level, which means each mipmap level only requires 1/4th as much CPU to draw as the previous level. So the whole mipmap drawing process of a 2048x2048 texture which would have taken 2760 milliseconds on an HTC One will instead take about 306 milliseconds.